requirements.txt
python-dotenv == 1.0.1
flask == 3.1.0 
flask_sqlalchemy == 3.1.1
flask-restful==0.3.10
psycopg2-binary==2.9.10